Finish Git Config

+++ src/main/java/com/gildo/springsecex/model/Users.java
@@ -6,17 +6,17 @@ import jakarta.persistence.Id;
 @Entity
 public class Users {
     @Id
-    private int id;
+    private Long id;
     private String username;
     private String password;
 
-    public Users(int id, String username, String password) {
+    public Users(Long id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
@@ -28,7 +28,7 @@ public class Users {
         return password;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
